bli-divorce/docs/existing/ReferenceCode/requirements.txt
asgiref==3.7.2
attrs==23.1.0
autopep8==2.0.2
Babel==2.12.1
boto3==1.26.145
botocore==1.29.145
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
cryptography==41.0.1
defusedxml==0.7.1
dictdiffer==0.9.0
dj-database-url==2.0.0
Django==4.2.1
django-axes==6.0.1
django-filter==23.2
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
django-storages==1.13.2
djangorestframework==3.14.0
docxcompose==1.4.0
docxtpl==0.16.7
easydict==1.10
et-xmlfile==1.1.0
fonttools==4.56.0
fpdf2==2.8.2
gunicorn==20.1.0
idna==3.4
inflect==6.0.4
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
lxml==4.9.2
MarkupSafe==2.1.2
openpyxl==3.1.2
phonenumbers==8.13.13
Pillow==9.5.0
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg-pool==3.1.7
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.8
pypdf==4.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
requests==2.32.1
s3transfer==0.6.1
sentry-sdk==1.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.4.0
